photo_ai/requirements.txt
fastapi==0.115.6
uvicorn==0.33.0
Pillow==10.4.0
numpy==1.24.4
opencv-python-headless==4.10.0.84
basicsr==1.4.2
realesrgan==0.3.0
